⌚ Fix command buffer wait and improve namings (#2)
Fixes a bug in destroying the command buffer pool, where command buffers
could still be in flight. Fixes a validation error.

Also as a drive by improves (temporary) command buffer naming, removes
the temp naming that was used, and removes the use of `shared_ptr` on a
`Fence`.

The latter is done because it doesn't make sense to have shared
ownership of a fence that is associated with a command buffer. It would
imply the `CommandBuffer` could be de-allocated prior to the `Fence`
being signaled, but this isn't allowed