🌏 Migrate to cmake (#6)
Significantly improves cross-platform-ability and solves the tons of
issues there were with setting up shader compilation embedded in the
build system.

As of this, this expects the CWD to be the build directory (textures and
shaders are accessible through this). Unfortunately, has to copy
textures at the moment, maybe find a way to symlink without being in
dev/admin mode