🤕 Fix some CMake issues (#7)
This fixes some CMake issues that popped up when using ninja from visual
studio. Missed an initial error and the build would result in cmake not
being able to load the cache right after a first build. The latter
happened because the custom step was deleting the directory, which is
obviously silly.

Also adds cmake presets to make this easier to run (particularly when
opened through non-VS generated solutions)